Fix WASM support.
